+++ src/lint/src/rules/feather/rules/index.ts
@@ -1,0 +1,80 @@
+export type { FeatherRuleFactory } from "../feather-rule-types.js";
+export {
+    createGm1013Rule,
+    createGm1032Rule,
+    createGm1034Rule,
+    createGm1036Rule,
+    createGm1056Rule,
+    createGm1059Rule,
+    createGm1062Rule
+} from "./gm-1xxx-extras.js";
+export {
+    createGm1000Rule,
+    createGm1002Rule,
+    createGm1003Rule,
+    createGm1004Rule,
+    createGm1005Rule,
+    createGm1007Rule,
+    createGm1008Rule,
+    createGm1009Rule,
+    createGm1010Rule,
+    createGm1012Rule,
+    createGm1014Rule,
+    createGm1015Rule,
+    createGm1016Rule,
+    createGm1017Rule,
+    createGm1021Rule,
+    createGm1023Rule,
+    createGm1024Rule,
+    createGm1026Rule,
+    createGm1028Rule,
+    createGm1029Rule,
+    createGm1030Rule,
+    createGm1033Rule,
+    createGm1038Rule,
+    createGm1041Rule,
+    createGm1051Rule,
+    createGm1052Rule,
+    createGm1054Rule,
+    createGm1058Rule,
+    createGm1063Rule,
+    createGm1064Rule,
+    createGm1100Rule
+} from "./gm-1xxx-rules.js";
+export {
+    createGm2000Rule,
+    createGm2003Rule,
+    createGm2004Rule,
+    createGm2005Rule,
+    createGm2007Rule,
+    createGm2008Rule,
+    createGm2009Rule,
+    createGm2011Rule,
+    createGm2012Rule,
+    createGm2015Rule,
+    createGm2020Rule,
+    createGm2023Rule,
+    createGm2025Rule,
+    createGm2026Rule,
+    createGm2028Rule,
+    createGm2029Rule,
+    createGm2030Rule,
+    createGm2031Rule,
+    createGm2032Rule,
+    createGm2033Rule,
+    createGm2035Rule,
+    createGm2040Rule,
+    createGm2042Rule,
+    createGm2043Rule,
+    createGm2044Rule,
+    createGm2046Rule,
+    createGm2048Rule,
+    createGm2050Rule,
+    createGm2051Rule,
+    createGm2052Rule,
+    createGm2053Rule,
+    createGm2054Rule,
+    createGm2056Rule,
+    createGm2061Rule,
+    createGm2064Rule
+} from "./gm-2xxx-rules.js";
